feat: add schedule utility helpers
